fix: remove package version from colophon (#202)
in #201, we introduced semantic-release, which manages and communicates
application version through git tags, and only updates `package.json`
upon release (usually distribution to `npm`).

Since this leaves the version string in `package.json` as the
placeholder for most cases in-repo, it's not appropriate to communicate
that string to users any more.

Will follow up later with a better mechanism to replace this.

Relates to #201