Add net_zero_reached_year variable to the schema (#216)
First of the variables required by the data model draft provided by
Climate Finance team added to the schema. The field is not required so
it is omitted from the dummy data for now. The variable is specified as
a `integer` with a minimum and maximum value as per Data model or `null`
to make it robust.

NOTE: early implementation was setting the type to "string", however as
per @AlexAxthelm feedback the "integer" was chosen in the end. See
https://github.com/RMI/pbtar/pull/216#issuecomment-3101789227 for the
discussion on string/integer and subsequent comments from @AlexAxthelm
for the full discussion. Other year variables in the schema should
follow suit in another PR.

---------

Co-authored-by: Alex Axthelm <alex@alexaxthelm.com>