ci(fix): warn, not error in GH for non-release to main
Fixes a GH Actions build error:
https://github.com/RMI/pbtar/actions/runs/17265869860/job/48997691047

where, since the last merge to `main` didn't trigger
`@semantic-release/exec`, there was no `dist.tgz` for the next step to
run.

This PR downgrades that from an error to a warning, on the logic that we
generally push a lot more features/fixes/etc. than chores (and generally
not chores consecutively), and anything in this branch that causes an
issue would get picked up in the next trigger of a versioned release.