fix: use `sticky`, not `fixed` for `EnvironmentBanner`
`App.tsx` was conditionally holding space for an environment banner, but the check
there was not sanitized in the same way as in `EnvironmentBanner.tsx`,
so under certain conditions (`production\n\n`), ther would be an empty
white space with no banner.

This moves all banner logic out of `App.tsx`, and holds space using
`sticky` class, rather than `fixed` and holding space.

Closes #330