ci(release): allow runs to not produce artifact
If `semantic-release` does not produce a new version for the workflow
run, then no `dist.tgz` build artifact is produced. This commit adds an
output flagging if that file is produced, so that the deploy step
doesn't try to run on a non-existent file.

The same logic applies in the PR arm of the `if`, to check for the
new release version, and falling back on the current version if not
found.

Closes #401