fix(colophon): colophon shows correct app version number
removing the `@semantic-release/npm` block from package.json in #635
(merge commit `6919803`) eliminated the mechanism by which the version
was updated in `package.json` (and then read by the app).

This change allows setting the version using an ENVVAR, and then setting
that during the `exec` phase when preparing the site tarball.

Closes #680